This PR optimizes out section B bottleneck (described here: https://github.com/zkSNACKs/WalletWasabi/issues/11287) of transaction processing with no noticeable memory growth.

This PR is a part of https://github.com/zkSNACKs/WalletWasabi/pull/11298 and https://github.com/zkSNACKs/WalletWasabi/pull/11513/

### master:

```
1.6GB memory
2023-09-23 10:51:43.331 [17] INFO	Wallet.Dispose (421)	Initial Transaction Processing finished in 9.2 seconds.
Time spent in sections: A: 0.42%, B: 24.02%, C: 1.40%, D: 28.57%, E: 7.25%, F: 1.05%, G: 37.28%
```

### This PR:

```
1.6 GB memory
2023-09-23 10:57:18.962 [20] INFO	Wallet.Dispose (421)	Initial Transaction Processing finished in 8.32 seconds.
2023-09-23 10:58:58.438 [17] INFO	Wallet.Dispose (421)	Initial Transaction Processing finished in 8.71 seconds.
2023-09-23 11:01:29.075 [28] INFO	Wallet.Dispose (421)	Initial Transaction Processing finished in 9.14 seconds.
2023-09-23 11:06:51.018 [9] INFO	Wallet.Dispose (421)	Initial Transaction Processing finished in 7.53 seconds.
Time spent in sections: A: 0.51%, B: 0.04%, C: 1.52%, D: 43.38%, E: 8.36%, F: 1.01%, G: 45.19%
```